+++ tests/Classes/Database/ConnectionTest.php
@@ -65,6 +65,7 @@ class ConnectionTest extends BaseTestCase
      */
     protected function prepareTsfeSetUp()
     {
+        $this->markTestSkipped('setup env with testing framework');
         $this->loadHiddenObjectsBackUp = $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['rn_base']['loadHiddenObjects'];
         $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['rn_base']['loadHiddenObjects'] = 0;
 
@@ -131,7 +132,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -154,7 +155,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -173,7 +174,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -185,7 +186,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -209,7 +210,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -232,7 +233,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -255,7 +256,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -275,7 +276,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -302,7 +303,7 @@ class ConnectionTest extends BaseTestCase
         $this->assertEquals($expected, $ret);
     }
 
-    public function singleFieldWhereProvider()
+    public static function singleFieldWhereProvider()
     {
         return [
             [OP_LIKE, 'm', ' '], // warum müssen mindestens 3 buchstaben vorliegen?
@@ -318,7 +319,7 @@ class ConnectionTest extends BaseTestCase
     }
 
     /**
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */
@@ -357,7 +358,7 @@ class ConnectionTest extends BaseTestCase
     /**
      * Tests the lookupLanguage method.
      *
-     * @group functional
+     * @group functional-old
      *
      * @TODO: refactor, requires tx_rnbase_util_TYPO3::getTSFE() which requires initialized database connection class
      */